StoreId contains upgrade info and is broadcasted to the whole cluster, slave checks all StoreIds on startup

+++ enterprise/com/src/main/java/org/neo4j/com/MismatchingVersionHandler.java
@@ -21,5 +21,5 @@ package org.neo4j.com;
 
 public interface MismatchingVersionHandler
 {
-    public void versionMismatched( int expected, int received );
+    public void versionMismatched( byte expected, byte received );
 }
